feat(cdk): enables customization of the API domain
While the Next.js rewrite support works nicely, it increases the total
cost of the deployment since the server API will be charged while the
real backend does its work. This gets worse when the backend is also
deployed as a lambda function on the same account, as they will consume
two concurrent execution quotas.